refactor: Move firebaseConfig file inside src
